+++ changelog.md
@@ -17,6 +17,7 @@
   - `SingleTableTransactionConfig` type renamed to `SingleTableTransactionParams`
 
 - *Feature*: Direct `collection` creation from `partition.collection()`
+- *Feature*: `propertiesToRetrieve` added to query methods
 - *Feature*: `schema.from(xxx).update()` now property infers return type if `returnUpdatedProperties` is true.
 - *Feature*: Entities' **Range Queries** now support definitions with only the operation provided. Required params will be the default for its respective operation